pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voice-dictation"
version = "0.1.0"
description = "Voice-to-text dictation tool using OpenAI Whisper API"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "sounddevice",
    "numpy",
    "pystray",
    "Pillow",
    "keyboard",
    "python-dotenv",
]

[project.scripts]
voice-dictation = "voice_dictation.app:main"

[tool.setuptools.packages.find]
where = ["src"]
